requirements.txt
# Core dependencies
chromadb==0.4.22
sentence-transformers==2.2.2
transformers==4.36.0
rank-bm25==0.2.2
networkx==3.1
numpy
pandas
scikit-learn

# API clients
groq==0.4.1
python-dotenv

# NLP
langdetect
nltk

# Utilities
tqdm
jupyter
ipython

deep-translator>=1.11,<2

langchain>=0.2.11,<0.3
langchain-community>=0.2.11,<0.3
langchain-huggingface>=0.0.3
